now it compresses and decodes .exe files

+++ Huffman/ByteFrequency.h
@@ -11,20 +11,20 @@ using namespace std;
 class ByteFrequency
 {
     private:
-        unsigned short int frequency;
+        unsigned int frequency;
         byte byte_code;
 
     public:
-        unsigned short int get_frequency() const;
+        unsigned int get_frequency() const;
         byte get_byte_code() const;
 
-        void set_frequency(unsigned short int f);
+        void set_frequency(unsigned int f);
         void set_byte_code(byte b);
         void increment_frequency();
 
         bool is_empty();
 
-        ByteFrequency(unsigned short int f, byte b);
+        ByteFrequency(unsigned int f, byte b);
         ByteFrequency();
 
         friend ofstream& operator<< (ofstream& os, const ByteFrequency& bf);
